User/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
bottle==0.13.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
gmssl==3.2.2
idna==3.10
loguru==0.7.3
packaging==25.0
pefile==2023.2.7
pip==25.0.1
proxy-tools==0.1.0
pyOpenSSL==25.0.0
pycparser==2.22
pycryptodome==3.22.0
pycryptodomex==3.22.0
pydantic==2.11.3
pydantic-core==2.33.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pymongo==4.12.0
pythonnet==3.0.5
pywebview==5.4
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==68.2.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
wheel==0.41.2
win32-setctime==1.2.0
screeninfo==0.8.1
